+++ server/src/main.ts
@@ -1,5 +1,6 @@
 import { Connection, ProposedFeatures, createConnection } from 'vscode-languageserver/node';
 import { activate } from './main-shared';
+import { MivaScriptCompilerProvider } from './mv/miva-script-compiler-provider/miva-script-compiler-provider';
 import { WorkspaceSymbolProvider } from './mv/symbol-provider/symbol-provider';
 import { formatError } from './util/functions';
 
@@ -14,4 +15,4 @@ process.on('uncaughtException', ( e: any ) => {
 	console.error( formatError( `Unhandled exception`, e ) );
 });
 
-activate(connection, new WorkspaceSymbolProvider());
+activate(connection, new WorkspaceSymbolProvider(), new MivaScriptCompilerProvider());